@@ -5,13 +5,13 @@
     "keywords": ["nosql","database","key-value store"],
     "license": "MIT",
     "require": {
-        "php": "^7.1",
+        "php": "^8.1",
         "ext-spl": "*",
         "ext-SimpleXML": "*", 
         "ext-xml": "*"
     },
     "require-dev": {
-    	"lucinda/unit-testing": "^1.0"
+    	"lucinda/unit-testing": "^2.0"
     },
     "autoload": {
     	"psr-4": {
