@@ -10,7 +10,7 @@ interface FileUpdater
      * Attempts to perform a value update, if applicable
      *
      * @param mixed $json Existing decoded entry value
-     * @return bool Whether or not update was done.
+     * @return bool Whether update was done.
      */
-    public function update(&$json): bool;
+    public function update(mixed &$json): bool;
 }
