@@ -4,10 +4,10 @@ namespace Lucinda\DB;
 /**
  * Enum defining statuses for schemas on health checks
  */
-interface SchemaStatus
+enum SchemaStatus: int
 {
-    const ONLINE = 1; // schema is up and running
-    const OFFLINE = 2; // schema is down
-    const UNRESPONSIVE = 3; // schema is up, but not responsive
-    const OVERLOADED = 4; // schema is up, but overloaded (operations take too long to execute)
+    case ONLINE = 1; // schema is up and running
+    case OFFLINE = 2; // schema is down
+    case UNRESPONSIVE = 3; // schema is up, but not responsive
+    case OVERLOADED = 4; // schema is up, but overloaded (operations take too long to execute)
 }
